RF: Move initial.py into netsim/cli/initial/__init__.py
Using the trick pioneered in 'netlab validate' refactoring, clean up
the 'netlab initial' implementation
